requirements.txt
# MCP SDK - the core Model Context Protocol library
mcp>=1.0.0

# HTTP client for making API requests to OpenWeatherMap
httpx>=0.27.0

# Environment variable management for API keys
python-dotenv>=1.0.0

# Anthropic SDK for Claude API (required by client.py)
anthropic>=0.40.0

# Optional: For better error messages and debugging
rich>=13.0.0